@@ -12,65 +12,55 @@ end
 
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.12'
-
-gem 'supplejack_common', git: 'git@github.com:DigitalNZ/supplejack_common.git'
-
-gem "oai", git: "https://github.com/boost/ruby-oai.git"
-
-gem "active_model_serializers", git: "https://github.com/rails-api/active_model_serializers.git"
-
-gem "thin",         ">= 1.5.0"
-gem "mongoid",      "~> 3.1.3"
-gem "figaro",       ">= 0.5.0"
-gem "devise",       ">= 2.1.2"
-gem "kaminari",     "~> 0.14.1"
-
-gem "airbrake"
-
-gem 'state_machine'
-gem 'sidekiq'
-gem 'kiqstand'
-
+gem 'rails', '4.1.4'
+gem 'protected_attributes'
+gem 'supplejack_common', git: 'git@github.com:DigitalNZ/supplejack_common.git', branch: 'rails-mongo-upgrade'
+# gem 'supplejack_common', path: '/code/boost/supplejack/supplejack_common'
+gem 'oai', git: 'https://github.com/boost/ruby-oai.git'
+gem 'active_model_serializers', '~> 0.9.0'
+gem 'thin',         '>= 1.5.0'
+gem 'mongoid',      '~> 4.0.0'
+gem 'mongoid_paranoia'
+gem 'figaro',       '>= 0.7.0'
+gem 'devise',       '~> 3.0.4'
+gem 'kaminari',     '~> 0.14.1'
+gem 'airbrake'
+gem 'aasm',         '~> 3.3.3'
+gem 'sidekiq',      '~> 2.15.0'
 # if you require 'sinatra' you get the DSL extended to Object
 gem 'sinatra', :require => nil
-
 gem 'activeresource', require: 'active_resource'
-
 gem 'whenever', :require => false
 gem 'parse-cron'
-gem 'strong_parameters'
+gem 'jquery-rails'
 
 # Gems used only for assets and not required
 # in production environments by default.
 group :assets do
-  gem 'sass-rails',   '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
-
+  gem 'sass-rails',   '~> 4.0.3'
+  gem 'coffee-rails', '~> 4.0.1'
   # See https://github.com/sstephenson/execjs#readme for more supported runtimes
   gem 'therubyracer', :platforms => :ruby
-
   gem 'uglifier', '>= 1.0.3'
 end
 
-gem 'jquery-rails'
-
 group :development do
-  gem "quiet_assets",       ">= 1.0.1"
-  gem "better_errors",      ">= 0.2.0"
+  gem 'quiet_assets',       '>= 1.0.1'
+  gem 'better_errors',      '~> 1.1.0 '
   gem 'guard-rspec'
   gem 'rb-fsevent',         '~> 0.9.1'
   gem 'pry-rails'
 end
 
 group :development, :test do
-  gem "rspec-rails", "~> 2.14.0"
-  gem "factory_girl_rails", ">= 4.1.0"
+  gem 'rspec', '~> 2.14.0'
+  gem 'rspec-rails', '~> 2.14.0'
+  gem 'factory_girl_rails', '>= 4.1.0'
 end
 
 group :test do
-  gem "database_cleaner",   ">= 0.9.1"
-  gem "cucumber-rails",     ">= 1.3.0", :require => false
+  gem 'database_cleaner',   '~> 1.3.0'
+  gem 'cucumber-rails',     '>= 1.4.1', :require => false
   gem 'timecop'
-  gem "rspec-sidekiq"
+  gem 'rspec-sidekiq'
 end
