@@ -14,12 +14,15 @@ class ApiUpdateWorker
 		attributes.merge!(preview: true) if job.environment == "preview"
 
 		measure = Benchmark.measure do
+      Sidekiq.logger.info "POST: #{ENV["API_HOST"]}#{path}"
+      Sidekiq.logger.info "Parameters: #{attributes.to_json}"
+
       response = RestClient.post "#{ENV["API_HOST"]}#{path}", attributes.to_json, content_type: :json, accept: :json
       response = JSON.parse(response)
-      job.set(:last_posted_record_id, response["record_id"])
+      job.set(last_posted_record_id: response["record_id"])
     end
 
-    job.inc(:posted_records_count, 1)
+    job.inc(:posted_records_count => 1)
 
     Sidekiq.logger.info "POST #{job.class} #{job.environment.capitalize} to #{path}. Time: #{measure.real.round(4)}s"
 	end
