@@ -13,9 +13,6 @@ HarvesterWorker::Application.configure do
   # since you don't have to restart the web server when you make code changes.
   config.cache_classes = false
 
-  # Log error messages when you accidentally call methods on nil.
-  config.whiny_nils = true
-
   # Show full error reports and disable caching
   config.consider_all_requests_local       = true
   config.action_controller.perform_caching = false
@@ -36,6 +33,9 @@ HarvesterWorker::Application.configure do
 
   # Expands the lines which load the assets
   config.assets.debug = true
+
+  # Do not eager load code on boot.
+  config.eager_load = false
 
   config.action_mailer.delivery_method = :smtp
   
